added cat command
